Added code to pull out pvals
Be careful when working with intercept, have to explicitly remove
And it isn't named like a regular variable.  Still need to handle
labeling the vars correctly.  Basically, if the intercept exists,
it comes first, no matter way.  I don't want to rely on the user
passing it in as the first variable...maybe when building formula
I force it to the beginning?